get_current_time function added
