+++ openai_realtime_streamlit/utils.py
@@ -3,13 +3,11 @@ import base64
 import json
 import numpy as np
 import os
-import queue
 import tzlocal
 from datetime import datetime
 from inspect import signature, Parameter
 from typing import Dict, Any, List, Optional
 
-import sounddevice as sd
 import websockets
 
 
@@ -23,7 +21,7 @@ class SimpleRealtime:
         self.ws = None
         self._message_handler_task = None
         self.audio_buffer_cb = audio_buffer_cb
-        self.tools = {}
+        self.tools = {}  # Added for tool support
 
     def _function_to_schema(self, func: callable) -> Dict[str, Any]:
         """
@@ -280,44 +278,3 @@ class SimpleRealtime:
         return True
 
 
-class StreamingAudioRecorder:
-    """
-    Thanks Sonnet 3.5...
-    """
-
-    def __init__(self, sample_rate=24_000, channels=1):
-        self.sample_rate = sample_rate
-        self.channels = channels
-        self.audio_queue = queue.Queue()
-        self.is_recording = False
-        self.audio_thread = None
-
-    def callback(self, indata, frames, time, status):
-        """
-        This will be called for each audio block
-        that gets recorded.
-        """
-        self.audio_queue.put(indata.copy())
-
-    def start_recording(self):
-        self.is_recording = True
-        self.audio_thread = sd.InputStream(
-            dtype="int16",
-            samplerate=self.sample_rate,
-            channels=self.channels,
-            callback=self.callback,
-            blocksize=2_000
-        )
-        self.audio_thread.start()
-
-    def stop_recording(self):
-        if self.is_recording:
-            self.is_recording = False
-            self.audio_thread.stop()
-            self.audio_thread.close()
-
-    def get_audio_chunk(self):
-        try:
-            return self.audio_queue.get_nowait()
-        except queue.Empty:
-            return None
